+++ ItsGrey/data/drawing/src/main/java/com/tntt/data/drawing/model/DrawingDto.kt
@@ -3,8 +3,8 @@ package com.tntt.data.drawing.model
 data class DrawingDto(
     var id: String,
     val imageBoxId: String,
-    val penSizeList: List<Int>,
-    val eraserSizeList: List<Int>,
+    val penSizeList: List<Long>,
+    val eraserSizeList: List<Long>,
     val penColor: String,
     val recentColors: List<String>,
 )
