+++ ItsGrey/data/layer/src/main/java/com/tntt/layer/model/LayerDto.kt
@@ -5,6 +5,6 @@ import android.graphics.Bitmap
 data class LayerDto(
     var id: String,
     val imageBoxId: String,
-    val order: Int,
+    val order: Long,
     val bitmap: Bitmap,
 )
