+++ ItsGrey/domain/model/src/main/java/com/tntt/model/DrawingInfo.kt
@@ -2,8 +2,8 @@ package com.tntt.model
 
 data class DrawingInfo(
     var id: String,
-    val penSizeList: List<Int>,
-    val eraserSizeList: List<Int>,
+    val penSizeList: List<Long>,
+    val eraserSizeList: List<Long>,
     val penColor: String,
     val recentColorList: List<String>,
 )
