+++ ItsGrey/domain/model/src/main/java/com/tntt/model/PageInfo.kt
@@ -2,5 +2,5 @@ package com.tntt.model
 
 data class PageInfo(
     val id: String,
-    val order: Int,
+    val order: Long,
 )
