+++ ItsGrey/domain/repo/src/main/java/com/tntt/repo/PageRepository.kt
@@ -6,7 +6,7 @@ import kotlinx.coroutines.flow.Flow
 
 interface PageRepository {
     suspend fun createPageInfo(bookId: String, pageInfo: PageInfo): Flow<PageInfo>
-    suspend fun getPageInfo(bookId: String, pageOrder: Int): Flow<PageInfo>
+    suspend fun getPageInfo(bookId: String, pageOrder: Long): Flow<PageInfo>
     suspend fun getPageInfoList(bookId: String): Flow<List<PageInfo>>
     suspend fun getFirstPageInfo(bookId: String): Flow<PageInfo?>
     suspend fun updatePageInfoList(bookId: String, pageInfoList: List<PageInfo>): Flow<Boolean>
